Monitor: made categories as a bit mask - now several categories can be requested at once. Monitor: added logs on server events: new connection, request of statistics, response, exceptions. Monitor: added restarting io_service after exception. Pytests: made disabling backends at test_recovery parallel.

+++ monitor/io_stat_provider.hpp
@@ -31,7 +31,7 @@ public:
 	io_stat_provider(dnet_node *n): m_node(n) {}
 
 	virtual std::string json() const;
-	virtual bool check_category(int category) const;
+	virtual bool check_category(uint64_t category) const;
 
 private:
 	dnet_node *m_node;
